requirements.txt
fastapi
uvicorn #  Le serveur ASGI (Asynchronous Server Gateway Interface) nécessaire pour exécuter votre application FastAPI.
opencv-python
ultralytics
python-dotenv
websockets